♻️ We don't actually need the Input wrapper anymore
