--- libraries/core-react/src/components/TextField/Input/index.ts
@@ -1,1 +1,0 @@
-export { TextFieldInput } from './Input'
