+++ libraries/core-react/src/components/TextField/TextField.tsx
@@ -1,15 +1,24 @@
 import * as React from 'react'
 import { ReactNode, InputHTMLAttributes, TextareaHTMLAttributes } from 'react'
 import styled from 'styled-components'
-import { TextFieldInput } from './Input'
+import { Icon } from './Icon'
+import { Input } from '../Input'
 import { Label } from '../Label'
 import { HelperText } from './HelperText'
-import { TextFieldProvider } from './context'
+import { TextFieldProvider, useTextField } from './context'
 import type { Variants } from './types'
 
 const Container = styled.div`
   min-width: 100px;
   width: 100%;
+`
+
+const RelativeContainer = styled.div`
+  position: relative;
+`
+
+const PaddedInput = styled(Input)`
+  padding-right: 32px;
 `
 
 export type TextFieldProps = {
@@ -94,12 +103,32 @@ export const TextField = React.forwardRef<HTMLInputElement, TextFieldProps>(
 
     const showLabel = label || meta
     const showHelperText = helperText
+    const { handleFocus, handleBlur } = useTextField()
 
     return (
       <Container {...containerProps}>
         <TextFieldProvider>
           {showLabel && <Label {...labelProps} />}
-          <TextFieldInput {...inputProps} />
+          <RelativeContainer>
+            {inputIcon ? (
+              <>
+                <PaddedInput
+                  onBlur={handleBlur}
+                  onFocus={handleFocus}
+                  {...inputProps}
+                />
+                <Icon isDisabled={disabled} variant={variant}>
+                  {inputIcon}
+                </Icon>
+              </>
+            ) : (
+              <Input
+                onBlur={handleBlur}
+                onFocus={handleFocus}
+                {...inputProps}
+              />
+            )}
+          </RelativeContainer>
           {showHelperText && <HelperText {...helperProps} />}
         </TextFieldProvider>
       </Container>
